telegabot/requirements.txt
python-telegram-bot==20.5
requests
pytz
httpx==0.24.1
